Move fasd and oh-my-zsh installation to utilities and cleanup

+++ install.sh
@@ -6,26 +6,18 @@ cd
 homedir="`pwd`"
 cd "$scriptdir"
 
+function exists() {
+	if command -v "$1" >/dev/null 2>&1;
+	then
+		echo 1
+	else
+		echo 0
+	fi
+}
+
 echo "initializing submodules"
 git submodule init
 git submodule update
-
-echo "Downloading and installing oh-my-zsh"
-wget --quiet --no-check-certificate https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh -O - | sh
-
-
-if ! [ "$(exists fasd)" -eq 1 ]
-then
-	echo "Downloading fasd"
-	git clone --quiet https://github.com/clvv/fasd.git
-
-	echo "Setting up FASD - a command-line productivity booster"
-	cd fasd
-	make install --quiet
-	cd ..
-	rm -r fasd/
-fi
-
 
 # rclocal=$(cat <<'END_HEREDOC'
 # #!/bin/sh -e
@@ -46,18 +38,16 @@ fi
 # exit 0
 # END_HEREDOC
 # )
+if [ "$(exists i3)" -eq 1 ]
+then
+	echo "Setting up the screen brightness script"
+	unlink $homedir/.i3/brightness
+	ln -s $scriptdir/brightness.sh $homedir/.i3/brightness
+	# 	sudo mv /etc/rc.local /etc/rc.local_old
+	# 	sudo sed -i -e '$i chown pedro:users /sys/class/backlight/radeon_bl0/brightness\n' "/etc/rc.local"
+	# 	# bash -c "echo "$rclocal" > "/etc/rc.local""
+fi
 
-# echo "$rclocal"
-
-# if [ "$(exists i3)" -eq 1 ]
-# then
-echo "Setting up the screen brightness script"
-ln -s $scriptdir/brightness.sh $homedir/.i3/brightness
-
-# 	sudo mv /etc/rc.local /etc/rc.local_old
-# 	sudo sed -i -e '$i chown pedro:users /sys/class/backlight/radeon_bl0/brightness\n' "/etc/rc.local"
-# 	# bash -c "echo "$rclocal" > "/etc/rc.local""
-# fi
 
 echo "Deleting the old files"
 mv ~/.zshrc ~/.zshrc_old
@@ -78,8 +68,8 @@ ln -s $scriptdir/gitignore ~/.gitignore
 ln -s $scriptdir/tmux ~/.tmux.conf
 ln -s $scriptdir/gemrc ~/.gemrc
 ln -s $scriptdir/bash_profile ~/.bash_profile
-mkdir ~/.i3/
-mkdir ~/.config/i3status/
+mkdir ~/.i3/ 2>/dev/null
+mkdir ~/.config/i3status/ 2>/dev/null
 ln -s $scriptdir/i3config ~/.i3/config
 ln -s $scriptdir/i3status ~/.config/i3status/config
